// RMS-BackOfficePortal/src/app/core/models/branch-code-counter.interface.ts
export interface BranchCodeCounter {
    bcc_id: number;
    counter_id: string;
    terminal_id: string;
    counter_ip: string;
    bcm_id: number;
    bcm_code: string;
    dt_created: Date;
    dt_modified: Date;
    created_by: string;
    modified_by: string;
    status_en: string;
    status_bm: string;
}